Move resample_volume into volume_operations
